add LabelCard and LabelEdit
